+++ web/src/lib/components/comparison/comparison-toolbar.svelte
